app respaldo/requirements.txt
paho-mqtt
pyserial
PyYAML
requests
pydantic
RPi.GPIO
tb-mqtt-client